requirements.txt
django==5.2.6
Pillow>=10.0.0
markdown
django-filter
djangorestframework
python-decouple
psycopg2-binary
dj-database-url
whitenoise
djangorestframework-simplejwt
gunicorn==21.2.0
